Bind question/answer/comment objects to users
* Update SQL scripts so that they contain this information
* Make the SQL schema stricter
* Update "add" methods so that they save user_id values
* Add Users and User page

+++ AskMateWebApp/Controllers/CommentsController.cs
@@ -25,7 +25,7 @@ namespace AskMateWebApp.Controllers
             AddCommentModel model;
             if (comment.AnswerId > 0)
             {
-                var answer = _answersService.GetOne(comment.AnswerId);
+                var answer = _answersService.GetOne((int)comment.AnswerId);
                 model = new AddCommentModel(answer.QuestionId, comment.Message);
             }
             else
